--- a/core/list.h
+++ b/core/list.h
@@ -444,6 +444,16 @@ public:
 		}
 	}
 
+	void operator=(List &&p_list) {
+		if (unlikely(this == &p_list)) {
+			return;
+		}
+
+		clear();
+		_data = p_list._data;
+		p_list._data = nullptr;
+	}
+
 	T &operator[](int p_index) {
 		CRASH_BAD_INDEX(p_index, size());
 
@@ -683,6 +693,11 @@ public:
 			push_back(it->get());
 			it = it->next();
 		}
+	}
+
+	List(List &&p_list) {
+		_data = p_list._data;
+		p_list._data = nullptr;
 	}
 
 	List() {