@akien-mga 2025-03-21
> I'm not familiar with how different 3.x is to 4.x, but if the unit tests are passing it's unlikely there are any for containers to require differences in implementation.

Just noting that the unit tests in 3.x are very bare ones / practically inexistent so the coverage can't be trusted blindly. The unit tests suite in 4.x was a major effort for 4.0 and later which wasn't back ported to 3.x.

Not to say that those changes shouldn't be done in 3.x, but the real test will be when users try it.