@lawnjelly 2025-03-23
Thanks!

After having these explained in the core meeting, afaik due to the nature of these changes if they were to fail, they would likely fail straight away (even in the absence of rigorous unit tests in 3.x), and it should be pretty easy to revert if any problems. The more testing we get before the next official build the better imo.